requirements.txt
aiogram==3.13.1
APScheduler==3.10.4
supabase==2.10.0
python-dotenv==1.0.1
pytz==2024.1
